refactor so realtime and historical use same datum generator
